fix: move contConf back to run container
